pyproject.toml
[project]
name = "ig-scraper"
version = "0.1.0"
description = "Instagram scraping tool"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["instaloader>=4.15.1", "Pillow>=8.1.1", "python-dotenv>=1.2.0", "requests>=2.33.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # Pyflakes
    "W",    # pycodestyle warnings
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "SIM",  # flake8-simplify
    "PT",   # flake8-pytest-style
    "RET",  # flake8-return
    "PTH",  # flake8-use-pathlib
    "RUF",  # Ruff-specific rules
    "T20",  # flake8-print
    "S",    # flake8-bandit (security)
    "D",    # pydocstyle
    "A",    # flake8-builtins
    "TCH",  # flake8-type-checking
    "PIE",  # flake8-pie
    "PERF", # perflint
    "LOG",  # flake8-logging
]
ignore = [
    "E501", # Line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101", # assert allowed in tests
    "ARG",  # unused arguments common in fixtures
    "D101", # missing docstrings common in test classes
    "D102", # missing docstrings common in test methods
    "F401", # unused imports common in test files
]

"scripts/**/*.py" = [
    "T201",   # print is the intended output mechanism for scripts
    "PTH123", # open() is fine in scripts
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true

[tool.ruff.lint.isort]
case-sensitive = true
lines-after-imports = 2

[tool.mypy]
python_version = "3.12"
strict = false
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
disallow_incomplete_defs = true
ignore_missing_imports = false


[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
    "--import-mode=importlib",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests that hit real APIs",
]
filterwarnings = ["error"]


[tool.inline-snapshot]
shortcuts = { fix = ["create", "fix"], review = "review" }

[tool.coverage.run]
source = ["src/ig_scraper"]
branch = true
omit = ["*/tests/*", "*/__main__.py"]

[tool.coverage.report]
fail_under = 80
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    'if __name__ == "__main__"',
    "@overload",
]

[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "pytest-cov>=6.0",
    "pytest-mock>=3.14",
    "responses>=0.25",
    "mypy>=1.15",
    "ruff>=0.15.9",
    "ty>=0.0.29",
    "pip-audit>=2.8.0",
    "hypothesis>=6.100",
    "polyfactory>=2.0",
    "inline-snapshot>=0.20",
    "pytest-regressions>=2.0",
    "pytest-bdd>=8.0",
    "bandit>=1.7.0",  # Security linting
    "invoke>=2.0",  # Task automation
    "prospector[with_mypy,with_bandit]>=1.10",  # Comprehensive analysis
]

[tool.bandit]
exclude_dirs = ["tests", "scripts", ".venv", ".git"]
skips = ["B101"]  # Allow asserts in non-test code if needed

[tool.uv.pip]
generate-hashes = true

